refactor: remove router rpc
