+++ src/meta-srv/src/test_util.rs
@@ -14,12 +14,18 @@
 
 use std::sync::Arc;
 
-use common_meta::key::TableMetadataManager;
+use chrono::DateTime;
+use common_catalog::consts::{DEFAULT_CATALOG_NAME, DEFAULT_SCHEMA_NAME, MITO_ENGINE};
+use common_meta::key::{TableMetadataManager, TableMetadataManagerRef};
 use common_meta::peer::Peer;
 use common_meta::rpc::router::{Region, RegionRoute};
 use common_meta::sequence::Sequence;
 use common_meta::state_store::KvStateStore;
 use common_procedure::local::{LocalManager, ManagerConfig};
+use datatypes::data_type::ConcreteDataType;
+use datatypes::schema::{ColumnSchema, RawSchema};
+use table::metadata::{RawTableInfo, RawTableMeta, TableIdent, TableType};
+use table::requests::TableOptions;
 
 use crate::cluster::MetaPeerClientBuilder;
 use crate::handler::{HeartbeatMailbox, Pushers};
@@ -87,4 +93,61 @@ pub(crate) fn create_region_failover_manager() -> Arc<RegionFailoverManager> {
         Arc::new(MemLock::default()),
         Arc::new(TableMetadataManager::new(KvBackendAdapter::wrap(kv_store))),
     ))
+}
+
+pub(crate) async fn prepare_table_region_and_info_value(
+    table_metadata_manager: &TableMetadataManagerRef,
+    table: &str,
+) {
+    let table_info = RawTableInfo {
+        ident: TableIdent::new(1),
+        name: table.to_string(),
+        desc: None,
+        catalog_name: DEFAULT_CATALOG_NAME.to_string(),
+        schema_name: DEFAULT_SCHEMA_NAME.to_string(),
+        meta: RawTableMeta {
+            schema: RawSchema::new(vec![ColumnSchema::new(
+                "a",
+                ConcreteDataType::string_datatype(),
+                true,
+            )]),
+            primary_key_indices: vec![],
+            value_indices: vec![],
+            engine: MITO_ENGINE.to_string(),
+            next_column_id: 1,
+            region_numbers: vec![1, 2, 3, 4],
+            options: TableOptions::default(),
+            created_on: DateTime::default(),
+            partition_key_indices: vec![],
+        },
+        table_type: TableType::Base,
+    };
+
+    let region_route_factory = |region_id: u64, peer: u64| RegionRoute {
+        region: Region {
+            id: region_id.into(),
+            ..Default::default()
+        },
+        leader_peer: Some(Peer {
+            id: peer,
+            addr: String::new(),
+        }),
+        follower_peers: vec![],
+    };
+
+    // Region distribution:
+    // Datanode => Regions
+    // 1 => 1, 2
+    // 2 => 3
+    // 3 => 4
+    let region_routes = vec![
+        region_route_factory(1, 1),
+        region_route_factory(2, 1),
+        region_route_factory(3, 2),
+        region_route_factory(4, 3),
+    ];
+    table_metadata_manager
+        .create_table_metadata(table_info, region_routes)
+        .await
+        .unwrap();
 }
