+++ src/partition/src/manager.rs
@@ -21,6 +21,7 @@ use common_meta::kv_backend::KvBackendRef;
 use common_meta::peer::Peer;
 use common_meta::rpc::router::{convert_to_region_map, RegionRoutes};
 use common_query::prelude::Expr;
+use common_telemetry::timer;
 use datafusion_expr::{BinaryExpr, Expr as DfExpr, Operator};
 use datatypes::prelude::Value;
 use snafu::{ensure, OptionExt, ResultExt};
@@ -29,6 +30,7 @@ use table::metadata::TableId;
 
 use crate::columns::RangeColumnsPartitionRule;
 use crate::error::{FindLeaderSnafu, Result};
+use crate::metrics::METRIC_TABLE_ROUTE_GET;
 use crate::partition::{PartitionBound, PartitionDef, PartitionExpr};
 use crate::range::RangePartitionRule;
 use crate::splitter::RowSplitter;
@@ -66,6 +68,7 @@ impl PartitionRuleManager {
 
     /// Find table route of given table name.
     pub async fn find_table_route(&self, table_id: TableId) -> Result<RegionRoutes> {
+        let _timer = timer!(METRIC_TABLE_ROUTE_GET);
         let route = self
             .table_route_manager
             .get(table_id)
